backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.0
